added clear color to camera struct

+++ engine/src/performance/performance.h
@@ -4,4 +4,4 @@
 
 MAPI void time_it_begin();
 MAPI void time_it_end();
-MAPI f32 time_get_elapsed(); // milliseconds
+MAPI f64 time_get_elapsed(); // seconds
